requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.9
django-bootstrap-v5==1.0.7
django-pipeline==2.0.7
idna==3.3
libsass==0.21.0
libsasscompiler==0.1.9
mercadopago==2.0.8
psycopg2==2.9.2
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
django-inlinecss<0.4
django-templated-email<3.1
pillow
qrcode